requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
enum34==1.1.6
hyperlink==17.3.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
jieba==0.38
lxml==3.8.0
mysql==0.0.1
MySQL-python==1.2.5
parsel==1.2.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.4.0
pyOpenSSL==17.2.0
queuelib==1.4.2
readability-lxml==0.6.2
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
w3lib==1.18.0
zope.interface==4.4.2
